+++ src/main/kotlin/ru/foodbox/delivery/services/model/DeliveryInfo.kt
@@ -1,8 +1,6 @@
 package ru.foodbox.delivery.services.model
 
-import java.math.BigDecimal
-
 data class DeliveryInfo(
-    val deliveryPrice: BigDecimal,
-    val freeDeliveryPrice: BigDecimal?,
+    val deliveryPrice: Long,
+    val freeDeliveryPrice: Long?,
 )
